+++ library/src/schemas/enumType/index.ts
@@ -1,2 +1,3 @@
 export * from './enumType.ts';
 export * from './enumTypeAsync.ts';
+export * from './types.ts';