+++ library/src/schemas/enumType/types.ts
@@ -1,0 +1,6 @@
+/**
+ * Enum type.
+ */
+export type Enum<TOption extends string = string> =
+  | Readonly<[TOption, ...TOption[]]>
+  | [TOption, ...TOption[]];